+++ src/main/java/org/apache/commons/imaging/formats/tiff/TiffHeader.java
@@ -22,9 +22,9 @@ import org.apache.commons.imaging.formats.tiff.constants.TiffConstants;
 public class TiffHeader extends TiffElement {
     public final ByteOrder byteOrder;
     public final int tiffVersion;
-    public final int offsetToFirstIFD;
+    public final long offsetToFirstIFD;
 
-    public TiffHeader(final ByteOrder byteOrder, final int tiffVersion, final int offsetToFirstIFD) {
+    public TiffHeader(final ByteOrder byteOrder, final int tiffVersion, final long offsetToFirstIFD) {
         super(0, TiffConstants.TIFF_HEADER_SIZE);
 
         this.byteOrder = byteOrder;
